+++ packages/core/src/lib/utils/assert.ts
@@ -1,4 +1,4 @@
-import { defaultCookies } from "./cookie.js"
+import { defaultCookies } from "../cookie.js"
 import {
   InvalidCallbackUrl,
   InvalidEndpoints,
@@ -8,10 +8,10 @@ import {
   MissingSecret,
   UnsupportedStrategy,
   UntrustedHost,
-} from "../errors.js"
+} from "../../errors.js"
 
-import type { AuthConfig, RequestInternal } from "../types.js"
-import type { WarningCode } from "./utils/logger.js"
+import type { AuthConfig, RequestInternal } from "../../types.js"
+import type { WarningCode } from "./logger.js"
 
 type ConfigError =
   | InvalidCallbackUrl
