+++ packages/core/src/lib/utils/callback-url.ts
@@ -1,4 +1,4 @@
-import type { InternalOptions } from "../types.js"
+import type { InternalOptions } from "../../types.js"
 
 interface CreateCallbackUrlParams {
   options: InternalOptions
